Finished AES Encryption/Decryption
Things i learned:
AES key can be generated from the EEPROM.
AES IV is generated randomly from a method in the library.
The IV needs padding so that it can be the correct length.
The IV needs to be generated only once, if it was generated twice then it will fail to do the decryption.